move hook to www, and autopep8
