+++ master/buildbot/www/hooks/bitbucket.py
